Fix stitching bug on AMD cards.
Something is slightly off with floor() and mod() on AMD GPUs,
which causes floor(n*3/3) = n-1 and mod(n*3, 3) = 3 instead of
floor(n*3/3) = n and mod(n*3, 3) = 0 as expected on NVIDIA GPUs.
This leads to grid artifacts on the stitched image because tile
indices are off by one.
To fix this,
    mod( xg, spw )
is replaced by
    trunc( mod( xg + 0.1, spw ) )
(This works reliably because xg is always integral.)